+++ strucmotif-search-core/src/main/java/org/rcsb/strucmotif/domain/query/MotifQuery.java
@@ -8,7 +8,7 @@ import java.util.Set;
 /**
  * The immutable container for a structural motif query.
  */
-public class MotifSearchQuery implements SearchQuery<MotifParameters, MotifQueryStructure> {
+public class MotifQuery implements SearchQuery<MotifParameters, MotifQueryStructure> {
     private final MotifQueryStructure queryStructure;
     private final Set<EnrichedMotifDefinition> motifDefinitions;
     private final MotifParameters parameters;
@@ -20,10 +20,10 @@ public class MotifSearchQuery implements SearchQuery<MotifParameters, MotifQuery
      * @param motifDefinitions the motifs to consider
      * @param parameters query parameters
      */
-    public MotifSearchQuery(String structureIdentifier,
-                            Structure structure,
-                            Set<EnrichedMotifDefinition> motifDefinitions,
-                            MotifParameters parameters) {
+    public MotifQuery(String structureIdentifier,
+                      Structure structure,
+                      Set<EnrichedMotifDefinition> motifDefinitions,
+                      MotifParameters parameters) {
         this.queryStructure = new MotifQueryStructure(structureIdentifier, structure);
         this.motifDefinitions = motifDefinitions;
         this.parameters = parameters;